+++ fe2o3-amqp/src/session/builder.rs
@@ -98,6 +98,7 @@ cfg_transaction! {
                 control_link_acceptor: ControlLinkAcceptor,
                 local_state: SessionState,
                 connection_stop_reason: Arc<OnceLock<ConnectionStopReason>>,
+                max_frame_size: usize,
             ) -> TxnSession<Session> {
                 let txn_manager = TransactionManager::new(outgoing, control_link_acceptor);
                 let session = Session {
@@ -131,6 +132,7 @@ cfg_transaction! {
                     control,
                     session,
                     txn_manager,
+                    max_frame_size,
                 }
             }
         }
@@ -297,7 +299,8 @@ impl Builder {
             };
 
             #[cfg(not(all(feature = "transaction", feature = "acceptor")))]
-            let (engine_handle, outcome, session_stop_reason) = {
+            let (engine_handle, outcome, session_stop_reason, max_frame_size) = {
+                let max_frame_size = connection.max_frame_size();
                 let session = self.into_session(
                     outgoing_channel,
                     local_state,
@@ -314,14 +317,15 @@ impl Builder {
                 )
                 .await?
                 .spawn();
-                (handle, outcome, session_stop_reason)
+                (handle, outcome, session_stop_reason, max_frame_size)
             };
 
             #[cfg(all(feature = "transaction", feature = "acceptor"))]
-            let (engine_handle, outcome, session_stop_reason) = {
+            let (engine_handle, outcome, session_stop_reason, max_frame_size) = {
                 let mut this = self;
                 match this.control_link_acceptor.take() {
                     Some(control_link_acceptor) => {
+                        let max_frame_size = connection.max_frame_size();
                         let session = this.into_txn_session(
                             session_control_tx.clone(),
                             outgoing_tx.clone(),
@@ -329,6 +333,7 @@ impl Builder {
                             control_link_acceptor,
                             local_state,
                             connection.connection_stop_reason.clone(),
+                            max_frame_size,
                         );
                         let session_stop_reason = session.session_stop_reason().clone();
                         let (handle, outcome) = SessionEngine::begin_client_session(
@@ -341,9 +346,10 @@ impl Builder {
                         )
                         .await?
                         .spawn();
-                        (handle, outcome, session_stop_reason)
+                        (handle, outcome, session_stop_reason, max_frame_size)
                     }
                     None => {
+                        let max_frame_size = connection.max_frame_size();
                         let session = this.into_session(
                             outgoing_channel,
                             local_state,
@@ -360,7 +366,7 @@ impl Builder {
                         )
                         .await?
                         .spawn();
-                        (handle, outcome, session_stop_reason)
+                        (handle, outcome, session_stop_reason, max_frame_size)
                     }
                 }
             };
@@ -372,6 +378,7 @@ impl Builder {
                 outcome,
                 outgoing: outgoing_tx,
                 session_stop_reason,
+                max_frame_size,
                 link_listener: (),
             };
             Ok(handle)
@@ -413,7 +420,8 @@ impl Builder {
                 },
             };
 
-            let (engine_handle, outcome, session_stop_reason) = {
+            let (engine_handle, outcome, session_stop_reason, max_frame_size) = {
+                let max_frame_size = connection.max_frame_size();
                 let session = self.into_session(
                     outgoing_channel,
                     local_state,
@@ -430,7 +438,7 @@ impl Builder {
                 )
                 .await?
                 .spawn_on_local_set(local_set);
-                (handle, outcome, session_stop_reason)
+                (handle, outcome, session_stop_reason, max_frame_size)
             };
 
             let handle = SessionHandle {
@@ -440,6 +448,7 @@ impl Builder {
                 outcome,
                 outgoing: outgoing_tx,
                 session_stop_reason,
+                max_frame_size,
                 link_listener: (),
             };
             Ok(handle)
@@ -480,7 +489,8 @@ impl Builder {
                 },
             };
 
-            let (engine_handle, outcome, session_stop_reason) = {
+            let (engine_handle, outcome, session_stop_reason, max_frame_size) = {
+                let max_frame_size = connection.max_frame_size();
                 let session = self.into_session(
                     outgoing_channel,
                     local_state,
@@ -497,7 +507,7 @@ impl Builder {
                 )
                 .await?
                 .spawn_local();
-                (handle, outcome, session_stop_reason)
+                (handle, outcome, session_stop_reason, max_frame_size)
             };
 
             let handle = SessionHandle {
@@ -507,6 +517,7 @@ impl Builder {
                 outcome,
                 outgoing: outgoing_tx,
                 session_stop_reason,
+                max_frame_size,
                 link_listener: (),
             };
             Ok(handle)